requirements.txt
# Web Application Dependencies
Flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
python-engineio==4.8.0
eventlet==0.33.3

# Database and Authentication
pymongo==4.6.0
bcrypt==4.1.2
flask-login==0.6.3
python-dotenv==1.0.0
